Move molecular utility functions to a new module
